requirements.txt
Django>=5.0,<6.0
django-jazzmin>=3.0,<4.0
waitress>=3.0,<4.0
whitenoise>=6.8,<7.0
